Move settings into app folder
