+++ tests/test_transform.py
@@ -1,4 +1,4 @@
-from server import app
+from transform import app
 
 import unittest
 import glob
